+++ Libplanet.Explorer.Executable/Options.cs
@@ -94,8 +94,9 @@ namespace Libplanet.Explorer.Executable
         {
             get
             {
-                return Seeds?.Select(seed => $"{ByteUtil.Hex(seed.PublicKey.Format(true))}," +
-                                             $"{seed.EndPoint.Host},{seed.EndPoint.Port}");
+                return Seeds?.Select(seed =>
+                    $"{ByteUtil.Hex(seed.PublicKey.CompressedKeyBytes.ToArray())}," +
+                    $"{seed.EndPoint.Host},{seed.EndPoint.Port}");
             }
 
             set