+++ Libplanet.Net.Tests/Consensus/Context/ContextNonProposerTest.cs
@@ -36,16 +36,20 @@ namespace Libplanet.Net.Tests.Consensus.Context
 
             Context.ProduceMessage(
                 TestUtils.CreateConsensusPropose(
-                    block, TestUtils.PrivateKeys[1], round: 0));
+                    block, TestUtils.ConsensusPrivateKeys[1], round: 0));
 
             Context.ProduceMessage(
                 TestUtils.CreateConsensusPropose(
-                    block, TestUtils.PrivateKeys[2], round: 1));
+                    block, TestUtils.ConsensusPrivateKeys[2], round: 1));
 
             Context.ProduceMessage(
                 new ConsensusVote(
                     TestUtils.CreateVote(
-                        TestUtils.PrivateKeys[2], 1, 1, hash: block.Hash, flag: VoteFlag.Absent))
+                        TestUtils.ConsensusPrivateKeys[2],
+                        1,
+                        1,
+                        hash: block.Hash,
+                        flag: VoteFlag.Absent))
                 {
                     Remote = TestUtils.Peers[2],
                 });
@@ -82,12 +86,12 @@ namespace Libplanet.Net.Tests.Consensus.Context
             Context.Start();
 
             Context.ProduceMessage(
-                TestUtils.CreateConsensusPropose(block, TestUtils.PrivateKeys[1]));
+                TestUtils.CreateConsensusPropose(block, TestUtils.ConsensusPrivateKeys[1]));
 
             Context.ProduceMessage(
                 new ConsensusVote(
                     TestUtils.CreateVote(
-                        TestUtils.PrivateKeys[1], 1, 0, hash: block.Hash, VoteFlag.Absent))
+                        TestUtils.ConsensusPrivateKeys[1], 1, 0, hash: block.Hash, VoteFlag.Absent))
                 {
                     Remote = TestUtils.Peers[1],
                 });
@@ -95,7 +99,7 @@ namespace Libplanet.Net.Tests.Consensus.Context
             Context.ProduceMessage(
                 new ConsensusVote(
                     TestUtils.CreateVote(
-                        TestUtils.PrivateKeys[2], 1, 0, hash: block.Hash, VoteFlag.Absent))
+                        TestUtils.ConsensusPrivateKeys[2], 1, 0, hash: block.Hash, VoteFlag.Absent))
                 {
                     Remote = TestUtils.Peers[2],
                 });
@@ -103,7 +107,7 @@ namespace Libplanet.Net.Tests.Consensus.Context
             Context.ProduceMessage(
                 new ConsensusVote(
                     TestUtils.CreateVote(
-                        TestUtils.PrivateKeys[3], 1, 0, hash: block.Hash, VoteFlag.Absent))
+                        TestUtils.ConsensusPrivateKeys[3], 1, 0, hash: block.Hash, VoteFlag.Absent))
                 {
                     Remote = TestUtils.Peers[3],
                 });
@@ -148,12 +152,12 @@ namespace Libplanet.Net.Tests.Consensus.Context
             Context.Start();
 
             Context.ProduceMessage(
-                TestUtils.CreateConsensusPropose(block, TestUtils.PrivateKeys[1]));
+                TestUtils.CreateConsensusPropose(block, TestUtils.ConsensusPrivateKeys[1]));
 
             Context.ProduceMessage(
                 new ConsensusVote(
                     TestUtils.CreateVote(
-                        TestUtils.PrivateKeys[1], 1, 0, hash: null, VoteFlag.Absent))
+                        TestUtils.ConsensusPrivateKeys[1], 1, 0, hash: null, VoteFlag.Absent))
                 {
                     Remote = TestUtils.Peers[1],
                 });
@@ -161,7 +165,7 @@ namespace Libplanet.Net.Tests.Consensus.Context
             Context.ProduceMessage(
                 new ConsensusVote(
                     TestUtils.CreateVote(
-                        TestUtils.PrivateKeys[2], 1, 0, hash: null, VoteFlag.Absent))
+                        TestUtils.ConsensusPrivateKeys[2], 1, 0, hash: null, VoteFlag.Absent))
                 {
                     Remote = TestUtils.Peers[2],
                 });
@@ -169,7 +173,7 @@ namespace Libplanet.Net.Tests.Consensus.Context
             Context.ProduceMessage(
                 new ConsensusVote(
                     TestUtils.CreateVote(
-                        TestUtils.PrivateKeys[3], 1, 0, hash: null, VoteFlag.Absent))
+                        TestUtils.ConsensusPrivateKeys[3], 1, 0, hash: null, VoteFlag.Absent))
                 {
                     Remote = TestUtils.Peers[3],
                 });
@@ -196,16 +200,16 @@ namespace Libplanet.Net.Tests.Consensus.Context
 
             Context.ProduceMessage(
                 TestUtils.CreateConsensusPropose(
-                    block, TestUtils.PrivateKeys[1], round: 0));
+                    block, TestUtils.ConsensusPrivateKeys[1], round: 0));
 
             Context.ProduceMessage(
                 TestUtils.CreateConsensusPropose(
-                    block, TestUtils.PrivateKeys[2], round: 1));
+                    block, TestUtils.ConsensusPrivateKeys[2], round: 1));
 
             Context.ProduceMessage(
                 new ConsensusVote(
                     TestUtils.CreateVote(
-                        TestUtils.PrivateKeys[2], 1, 1, hash: null, flag: VoteFlag.Absent))
+                        TestUtils.ConsensusPrivateKeys[2], 1, 1, hash: null, flag: VoteFlag.Absent))
                 {
                     Remote = TestUtils.Peers[2],
                 });
@@ -260,23 +264,23 @@ namespace Libplanet.Net.Tests.Consensus.Context
             // Push round 0 and round 1 proposes.
             Context.ProduceMessage(
                 TestUtils.CreateConsensusPropose(
-                    block1, TestUtils.PrivateKeys[1], round: 0));
+                    block1, TestUtils.ConsensusPrivateKeys[1], round: 0));
             Context.ProduceMessage(
                 TestUtils.CreateConsensusPropose(
-                    block2, TestUtils.PrivateKeys[2], round: 1));
+                    block2, TestUtils.ConsensusPrivateKeys[2], round: 1));
 
             // Two additional votes should be enough to trigger prevote timeout timer.
             Context.ProduceMessage(
                 new ConsensusVote(
                     TestUtils.CreateVote(
-                        TestUtils.PrivateKeys[2], 1, 0, hash: null, flag: VoteFlag.Absent))
+                        TestUtils.ConsensusPrivateKeys[2], 1, 0, hash: null, flag: VoteFlag.Absent))
                 {
                     Remote = TestUtils.Peers[1],
                 });
             Context.ProduceMessage(
                 new ConsensusVote(
                     TestUtils.CreateVote(
-                        TestUtils.PrivateKeys[3], 1, 0, hash: null, flag: VoteFlag.Absent))
+                        TestUtils.ConsensusPrivateKeys[3], 1, 0, hash: null, flag: VoteFlag.Absent))
                 {
                     Remote = TestUtils.Peers[1],
                 });
@@ -285,14 +289,14 @@ namespace Libplanet.Net.Tests.Consensus.Context
             Context.ProduceMessage(
                 new ConsensusCommit(
                     TestUtils.CreateVote(
-                        TestUtils.PrivateKeys[2], 1, 0, hash: null, flag: VoteFlag.Absent))
+                        TestUtils.ConsensusPrivateKeys[2], 1, 0, hash: null, flag: VoteFlag.Absent))
                 {
                     Remote = TestUtils.Peers[1],
                 });
             Context.ProduceMessage(
                 new ConsensusCommit(
                     TestUtils.CreateVote(
-                        TestUtils.PrivateKeys[3], 1, 0, hash: null, flag: VoteFlag.Absent))
+                        TestUtils.ConsensusPrivateKeys[3], 1, 0, hash: null, flag: VoteFlag.Absent))
                 {
                     Remote = TestUtils.Peers[1],
                 });
@@ -321,12 +325,16 @@ namespace Libplanet.Net.Tests.Consensus.Context
 
             Context.ProduceMessage(
                 TestUtils.CreateConsensusPropose(
-                    block, TestUtils.PrivateKeys[1], round: 0));
+                    block, TestUtils.ConsensusPrivateKeys[1], round: 0));
 
             Context.ProduceMessage(
                 new ConsensusVote(
                     TestUtils.CreateVote(
-                        TestUtils.PrivateKeys[1], 1, 0, hash: block.Hash, flag: VoteFlag.Absent))
+                        TestUtils.ConsensusPrivateKeys[1],
+                        1,
+                        0,
+                        hash: block.Hash,
+                        flag: VoteFlag.Absent))
                 {
                     Remote = TestUtils.Peers[1],
                 });
@@ -334,7 +342,7 @@ namespace Libplanet.Net.Tests.Consensus.Context
             Context.ProduceMessage(
                 new ConsensusVote(
                     TestUtils.CreateVote(
-                        TestUtils.PrivateKeys[2], 1, 0, hash: null, flag: VoteFlag.Absent))
+                        TestUtils.ConsensusPrivateKeys[2], 1, 0, hash: null, flag: VoteFlag.Absent))
                 {
                     Remote = TestUtils.Peers[2],
                 });
@@ -342,7 +350,7 @@ namespace Libplanet.Net.Tests.Consensus.Context
             Context.ProduceMessage(
                 new ConsensusVote(
                     TestUtils.CreateVote(
-                        TestUtils.PrivateKeys[3], 1, 0, hash: null, flag: VoteFlag.Absent))
+                        TestUtils.ConsensusPrivateKeys[3], 1, 0, hash: null, flag: VoteFlag.Absent))
                 {
                     Remote = TestUtils.Peers[3],
                 });
@@ -364,12 +372,16 @@ namespace Libplanet.Net.Tests.Consensus.Context
 
             Context.ProduceMessage(
                 TestUtils.CreateConsensusPropose(
-                    block, TestUtils.PrivateKeys[1], round: 0));
+                    block, TestUtils.ConsensusPrivateKeys[1], round: 0));
 
             Context.ProduceMessage(
                 new ConsensusCommit(
                     TestUtils.CreateVote(
-                        TestUtils.PrivateKeys[1], 1, 0, hash: block.Hash, flag: VoteFlag.Commit))
+                        TestUtils.ConsensusPrivateKeys[1],
+                        1,
+                        0,
+                        hash: block.Hash,
+                        flag: VoteFlag.Commit))
                 {
                     Remote = TestUtils.Peers[1],
                 });
@@ -377,7 +389,7 @@ namespace Libplanet.Net.Tests.Consensus.Context
             Context.ProduceMessage(
                 new ConsensusCommit(
                     TestUtils.CreateVote(
-                        TestUtils.PrivateKeys[2], 1, 0, hash: null, flag: VoteFlag.Commit))
+                        TestUtils.ConsensusPrivateKeys[2], 1, 0, hash: null, flag: VoteFlag.Commit))
                 {
                     Remote = TestUtils.Peers[2],
                 });
@@ -385,7 +397,7 @@ namespace Libplanet.Net.Tests.Consensus.Context
             Context.ProduceMessage(
                 new ConsensusCommit(
                     TestUtils.CreateVote(
-                        TestUtils.PrivateKeys[3], 1, 0, hash: null, flag: VoteFlag.Commit))
+                        TestUtils.ConsensusPrivateKeys[3], 1, 0, hash: null, flag: VoteFlag.Commit))
                 {
                     Remote = TestUtils.Peers[3],
                 });