+++ Libplanet.Net.Tests/SwarmTest.Fixtures.cs
@@ -72,7 +72,6 @@ namespace Libplanet.Net.Tests
 
         private Swarm<DumbAction> CreateSwarm(
             PrivateKey privateKey = null,
-            PrivateKey consensusPrivateKey = null,
             AppProtocolVersion? appProtocolVersion = null,
             string host = null,
             int? listenPort = null,
@@ -142,7 +141,7 @@ namespace Libplanet.Net.Tests
                 {
                     ConsensusPeers = ImmutableList<BoundPeer>.Empty,
                     ConsensusPort = 0,
-                    ConsensusPrivateKey = new PrivateKey(),
+                    ConsensusPrivateKey = new BlsPrivateKey(),
                     ConsensusWorkers = 100,
                     TargetBlockInterval = TimeSpan.FromSeconds(10),
                 });