+++ Libplanet.Net/Transports/InvalidMessageSignatureException.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Linq;
 using System.Runtime.Serialization;
 using Libplanet.Crypto;
 using Libplanet.Net.Messages;
@@ -16,7 +17,7 @@ namespace Libplanet.Net.Transports
         internal InvalidMessageSignatureException(
             string message,
             BoundPeer peer,
-            PublicKey publicKey,
+            IPublicKey publicKey,
             byte[] messageToVerify,
             byte[] signature)
             : base(message)
@@ -40,7 +41,7 @@ namespace Libplanet.Net.Transports
 
         public BoundPeer Peer { get; private set; }
 
-        public PublicKey PublicKey { get; private set; }
+        public IPublicKey PublicKey { get; private set; }
 
         public byte[] MessageToVerify { get; private set; }
 
@@ -51,7 +52,7 @@ namespace Libplanet.Net.Transports
         {
             base.GetObjectData(info, context);
             info.AddValue(nameof(Peer), Peer);
-            info.AddValue(nameof(PublicKey), PublicKey.Format(true));
+            info.AddValue(nameof(PublicKey), PublicKey.KeyBytes.ToArray());
             info.AddValue(nameof(MessageToVerify), MessageToVerify);
             info.AddValue(nameof(Signature), Signature);
         }