+++ CHANGES.md
@@ -63,8 +63,7 @@ To be released.
  -  Added `Crypto.IPublicKey` inherited `Crypto.BlsPublicKey` class.
     [[#PBFT]]
  -  Added `Crypto.BlsSignature` class.  [[#PBFT]]
- -  `Crypto.PublicKey` is now inheriting `Crypto.IPrivateKey`.  [[#PBFT]]
- -  `Crypto.PrivateKey` is now inheriting `Crypto.IPublicKey`.  [[#PBFT]]
+ -  Added static `PublicKeyGetter` class.  [[#PBFT]]
  -  (Libplanet.Net) Added `IReactor` interface.  [[#PBFT]]
  -  (Libplanet.Net) Added `ConsensusReactor` class which inherits
     `IReactor` interface.  [[#PBFT]]
@@ -97,6 +96,30 @@ To be released.
  -  `PreEvaluationBlockHeader()` constructor became to throw
     `InvalidBlockLastCommitException` when its metadata's `LastCommit` is
     invalid.  [[#PBFT]]
+ -  `Crypto.PublicKey` is now inheriting `Crypto.IPrivateKey`.  [[#PBFT]]
+ -  `Crypto.PrivateKey` is now inheriting `Crypto.IPublicKey`.  [[#PBFT]]
+ -  `Address(PublicKey)` constructor is changed to
+    `Address(IPublicKey)`.  [[#PBFT]]
+ -  `AddressExtensions.ToAddress(PublicKey)` is changed to
+    `AddressExtensions.ToAddress(IPublicKey)`.  [[#PBFT]]
+ -  `AddressExtensions.ToAddress(PrivateKey)` is changed to
+    `AddressExtensions.ToAddress(IPrivateKey)`.  [[#PBFT]]
+ -  (Libplanet.Net) Property `BoundPeer.PublicKey` type is now `IPublicKey`. the
+    following constructors `BoundPeer(PublicKey, IPAddress?)` and
+    `BoundPeer(PublicKey)` are also affected by this changes.  [[#PBFT]]
+ -  (Libplanet.Net) Parameter type `PublicKey` in `BoundPeer(PublicKey,
+    DnsEndPoint)` and `BoundPeer(PublicKey, DnsEndPoint, IPAddress?)` is
+    now `IPublicKey`.  [[#PBFT]]
+ -  (Libplanet.Net) Parameters type `PublicKey expected` and `PublicKey actual`
+    in `InvalidCredentialException` is now `IPublicKey`.  [[#PBFT]]
+ -  (Libplanet.Net) Parameter type `PublicKey publicKey` in
+    `InvalidMessageSignatureException` is now `IPublicKey`.  [[#PBFT]]
+ -  (Libplanet.Net) Parameter type `PrivateKey privateKey` in
+    `IMessageCodec<T>.Encode(Message, PrivateKey, AppProtocolVersion, Peer,
+    DateTimeOffset)` is now `IPrivatekey`.  [[#PBFT]]
+ -  (Libplanet.Net) Parameter type `PrivateKey privateKey` in
+    `NetMQMessageCodec.Encode(Message, PrivateKey, AppProtocolVersion, Peer,
+    DateTimeOffset)` is now `IPrivateKey`.  [[#PBFT]]
 
 ### Bug fixes